backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# LangChain & LangGraph
langchain==0.1.0
langchain-core==0.1.3
langchain-community==0.0.10
langgraph==0.0.25
langchain-google-genai==0.0.13
langchain-groq==0.0.5

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1

# Algorand
py-algorand-sdk==2.5.0
pyteal==0.20.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.1

# HTTP
httpx==0.25.2
aiofiles==23.2.1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging & Monitoring
loguru==0.7.2

# Math & Science
numpy==1.26.2
scipy==1.11.4

# Development
black==23.12.0
flake8==6.1.0

